[dev] - Updated home page presentation.
