+++ lib/app/modules/home_pokedex/ui/home_page/controller/pokedex_home_controller.dart
@@ -1,9 +1,9 @@
-import 'package:flutter_pokedex/app/modules/home_pokedex/domain/entity/pokemon_list_entity.dart';
 import 'package:mobx/mobx.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_pokedex/models/pokeApi.dart';
 import 'package:flutter_pokedex/app/constants/consts_app.dart';
 import 'package:cached_network_image/cached_network_image.dart';
+import 'package:flutter_pokedex/app/modules/home_pokedex/domain/entity/pokemon_entity.dart';
+import 'package:flutter_pokedex/app/modules/home_pokedex/domain/entity/pokemon_list_entity.dart';
 import 'package:flutter_pokedex/app/modules/home_pokedex/domain/usecase/get_list_pokemon_usecase.dart';
 part 'pokedex_home_controller.g.dart';
 
@@ -18,13 +18,16 @@ abstract class _PokedexHomeControllerBase with Store {
   }) : _getListPokemonUsecase = getListPokemonUsecase;
 
   @observable
-  PokeApi _pokeAPI;
-
-  @observable
   PokemonListEntity _pokemonList;
 
+  @computed
+  PokemonListEntity get pokemonList => _pokemonList;
+
   @observable
-  Pokemon _currentPokemon;
+  PokemonEntity _currentPokemon;
+
+  @computed
+  PokemonEntity get currentPokemon => _currentPokemon;
 
   @observable
   Color pokeColor;
@@ -32,30 +35,19 @@ abstract class _PokedexHomeControllerBase with Store {
   @observable
   int currentPosition;
 
-  @computed
-  PokeApi get pokeAPI => _pokeAPI;
-
-  @computed
-  Pokemon get currentPokemon => _currentPokemon;
-
   @action
   fetchPokemonList() async {
-    _pokeAPI = null;
+    _pokemonList = null;
     _pokemonList = await _getListPokemonUsecase.getPokemonList();
-    print(_pokemonList);
-
-    // loadPokeAPI().then((pokeList) {
-    //   _pokeAPI = pokeList;
-    // });
   }
 
-  Pokemon getPokemon({int index}) {
-    return _pokeAPI.pokemon[index];
+  PokemonEntity getPokemon({int index}) {
+    return _pokemonList.pokemonListEntity[index];
   }
 
   @action
   setCurrentPokemon({int index}) {
-    _currentPokemon = _pokeAPI.pokemon[index];
+    _currentPokemon = _pokemonList.pokemonListEntity[index];
     pokeColor = ConstsApp.getColorType(type: _currentPokemon.type[0]);
     currentPosition = index;
   }
