--- lib/app/modules/home_pokedex/ui/home_page/widgets/poke_item.dart
@@ -1,128 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_pokedex/app/constants/consts_app.dart';
-import 'package:cached_network_image/cached_network_image.dart';
-
-class PokeItem extends StatelessWidget {
-  final String name;
-  final int index;
-  final Color color;
-  final String num;
-  final List<String> types;
-
-  Widget setTipos() {
-    List<Widget> lista = [];
-    types.forEach((nome) {
-      lista.add(
-        Column(
-          children: <Widget>[
-            Container(
-              padding: EdgeInsets.all(0),
-              decoration: BoxDecoration(
-                  borderRadius: BorderRadius.circular(20),
-                  color: Color.fromARGB(80, 255, 255, 255)),
-              child: Padding(
-                padding: const EdgeInsets.all(6.0),
-                child: Text(
-                  nome.trim(),
-                  style: TextStyle(
-                      fontFamily: 'Google',
-                      fontSize: 16,
-                      fontWeight: FontWeight.bold,
-                      color: Colors.white),
-                ),
-              ),
-            ),
-            SizedBox(
-              height: 5,
-            )
-          ],
-        ),
-      );
-    });
-    return Column(
-      children: lista,
-      crossAxisAlignment: CrossAxisAlignment.start,
-    );
-  }
-
-  const PokeItem(
-      {Key key, this.name, this.index, this.color, this.num, this.types});
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(8.0),
-      child: Container(
-        child: Padding(
-          padding: const EdgeInsets.all(8.0),
-          child: Stack(
-            children: <Widget>[
-              Align(
-                alignment: Alignment.bottomRight,
-                child: Hero(
-                  child: Opacity(
-                    child: Image.asset(
-                      ConstsApp.whitePokeball,
-                      height: 130,
-                      width: 130,
-                    ),
-                    opacity: 0.2,
-                  ),
-                  tag: name + 'rotation',
-                ),
-              ),
-              Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: <Widget>[
-                  Padding(
-                    padding: const EdgeInsets.only(left: 4.0, top: 10),
-                    child: Text(
-                      name,
-                      style: TextStyle(
-                        fontFamily: 'Google',
-                        fontSize: 20.0,
-                        fontWeight: FontWeight.bold,
-                        color: Colors.white,
-                      ),
-                    ),
-                  ),
-                  Padding(
-                    padding: const EdgeInsets.all(4.0),
-                    child: setTipos(),
-                  ),
-                ],
-              ),
-              Align(
-                alignment: Alignment.bottomRight,
-                child: Hero(
-                  child: CachedNetworkImage(
-                    height: 80,
-                    width: 80,
-                    placeholder: (context, url) => new Container(
-                      color: Colors.transparent,
-                    ),
-                    imageUrl:
-                        'https://raw.githubusercontent.com/fanzeyi/pokemon.json/master/images/$num.png',
-                  ),
-                  tag: name,
-                ),
-              ),
-            ],
-          ),
-        ),
-        decoration: BoxDecoration(
-          gradient: LinearGradient(
-              begin: Alignment.topCenter,
-              end: Alignment.bottomCenter,
-              colors: [
-                ConstsApp.getColorType(type: types[0]),
-                ConstsApp.getColorType(type: types[0]).withOpacity(0.7),
-              ]),
-          borderRadius: BorderRadius.all(
-            Radius.circular(25.0),
-          ),
-        ),
-      ),
-    );
-  }
-}
